domain-api/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-Levenshtein
